Change mapmusrng to UINT8

+++ src/doomstat.h
@@ -42,7 +42,7 @@ extern char mapmusname[7];
 extern UINT16 mapmusflags;
 extern UINT32 mapmusposition;
 extern UINT32 mapmusresume;
-extern UINT32 mapmusrng;
+extern UINT8 mapmusrng;
 #define MUSIC_TRACKMASK   0x0FFF // ----************
 #define MUSIC_RELOADRESET 0x8000 // *---------------
 #define MUSIC_FORCERESET  0x4000 // -*--------------
